+++ cardano-node/src/Cardano/Node/Configuration/POM.hs
@@ -123,7 +123,7 @@ data NodeConfiguration
          --
          -- This flag should be set to 'True' when testing the new protocol
          -- versions.
-       , ncTestEnableDevelopmentNetworkProtocols :: !Bool
+       , ncExperimentalProtocols :: !Bool
 
          -- BlockFetch configuration
        , ncMaxConcurrencyBulkSync :: !(Maybe MaxConcurrencyBulkSync)
@@ -179,7 +179,7 @@ data PartialNodeConfiguration
          -- Node parameters, not protocol-specific:
        , pncDiffusionMode    :: !(Last DiffusionMode)
        , pncSnapshotInterval :: !(Last SnapshotInterval)
-       , pncTestEnableDevelopmentNetworkProtocols :: !(Last Bool)
+       , pncExperimentalProtocols :: !(Last Bool)
 
          -- BlockFetch configuration
        , pncMaxConcurrencyBulkSync :: !(Last MaxConcurrencyBulkSync)
@@ -230,8 +230,15 @@ instance FromJSON PartialNodeConfiguration where
         <- Last . fmap getDiffusionMode <$> v .:? "DiffusionMode"
       pncSnapshotInterval
         <- Last . fmap RequestedSnapshotInterval <$> v .:? "SnapshotInterval"
-      pncTestEnableDevelopmentNetworkProtocols
-        <- Last <$> v .:? "TestEnableDevelopmentNetworkProtocols"
+      pncExperimentalProtocols <- fmap Last $ do
+        failOnRemovedFields v
+          [ RemovedField
+              { rfField = "TestEnableDevelopmentNetworkProtocols"
+              , rfErr = "TestEnableDevelopmentNetworkProtocols has been renamed to ExperimentalProtocols"
+              }
+          ]
+
+        v .:? "ExperimentalProtocols"
 
       -- Blockfetch parameters
       pncMaxConcurrencyBulkSync <- Last <$> v .:? "MaxConcurrencyBulkSync"
@@ -295,7 +302,7 @@ instance FromJSON PartialNodeConfiguration where
            , pncSocketConfig = Last . Just $ SocketConfig mempty mempty mempty pncSocketPath
            , pncDiffusionMode
            , pncSnapshotInterval
-           , pncTestEnableDevelopmentNetworkProtocols
+           , pncExperimentalProtocols
            , pncMaxConcurrencyBulkSync
            , pncMaxConcurrencyDeadline
            , pncLoggingSwitch = Last $ Just pncLoggingSwitch'
@@ -464,7 +471,7 @@ defaultPartialNodeConfiguration =
     , pncSocketConfig = Last . Just $ SocketConfig mempty mempty mempty mempty
     , pncDiffusionMode = Last $ Just InitiatorAndResponderDiffusionMode
     , pncSnapshotInterval = Last $ Just DefaultSnapshotInterval
-    , pncTestEnableDevelopmentNetworkProtocols = Last $ Just False
+    , pncExperimentalProtocols = Last $ Just False
     , pncTopologyFile = Last . Just $ TopologyFile "configuration/cardano/mainnet-topology.json"
     , pncProtocolFiles = mempty
     , pncValidateDB = mempty
@@ -537,9 +544,9 @@ makeNodeConfiguration pnc = do
     $ pncEnableP2P pnc
 
   -- TODO: This is not mandatory
-  testEnableDevelopmentNetworkProtocols <-
-    lastToEither "Missing TestEnableDevelopmentNetworkProtocols" $
-      pncTestEnableDevelopmentNetworkProtocols pnc
+  experimentalProtocols <-
+    lastToEither "Missing ExperimentalProtocols" $
+      pncExperimentalProtocols pnc
   return $ NodeConfiguration
              { ncConfigFile = configFile
              , ncTopologyFile = topologyFile
@@ -557,7 +564,7 @@ makeNodeConfiguration pnc = do
              , ncSocketConfig = socketConfig
              , ncDiffusionMode = diffusionMode
              , ncSnapshotInterval = snapshotInterval
-             , ncTestEnableDevelopmentNetworkProtocols = testEnableDevelopmentNetworkProtocols
+             , ncExperimentalProtocols = experimentalProtocols
              , ncMaxConcurrencyBulkSync = getLast $ pncMaxConcurrencyBulkSync pnc
              , ncMaxConcurrencyDeadline = getLast $ pncMaxConcurrencyDeadline pnc
              , ncLoggingSwitch = loggingSwitch